+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ### Changes since the last release
 
+ * Greenhouses now act like a scrubber and also will not use CO2 or produce O2 when in a breathable atmosphere (PiezPiedPy)
  * Reliability: mass and cost difference between standard and high quality is now relative to the part type (Gordon Dry)
  * Allow vessel config when there is no vessel signal (PiezPiedPy)
  * Fixed EVA Scrubber, ooops was broken by changes to Habitation (PiezPiedPy)